requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pillow==10.0.0
opencv-python-headless==4.7.0.72
pytesseract==0.3.10
paddlepaddle==2.6.0
paddleocr>=2.7.0.1
numpy==1.24.3
python-magic==0.4.27
python-dotenv==1.0.0
pydantic==1.10.7
typing-extensions>=4.5.0
scikit-learn==1.3.2
joblib==1.3.2
